fix example
